Clean up save function, commit a few crimes
